basic sidekiq functionality

+++ app/services/marty/promises/cancel.rb
@@ -2,14 +2,40 @@ module Marty::Promises
   module Cancel
     class << self
       def call(id)
+        if Marty::Config['USE_SIDEKIQ_WITH_PROMISES'].to_s == 'true'
+          return call_with_sidekiq(id)
+        end
+
         ids = get_all_ids(id)
         promises = Marty::Promise.where(id: ids)
         jobids = promises.map(&:job_id).compact.sort
         Delayed::Job.where(id: jobids).destroy_all
         promises.each do |p|
-          p.update!(status: false,
-                    end_dt: p.end_dt || Time.zone.now,
-                    result: p.result + { error: 'Cancelled' })
+          p.update!(
+            status: false,
+            end_dt: p.end_dt || Time.zone.now,
+            result: p.result + { error: 'Cancelled' }
+          )
+        end
+      end
+
+      def call_with_sidekiq(id)
+        ids = get_all_ids(id)
+        promises = Marty::Promise.where(id: ids)
+        jobids = promises.map(&:job_id).compact.sort
+
+        Sidekiq::Queue.all.each do |queue|
+          queue.each do |job|
+            job.delete if jobids.include?(job.jid)
+          end
+        end
+
+        promises.each do |p|
+          p.update!(
+            status: false,
+            end_dt: p.end_dt || Time.zone.now,
+            result: p.result + { error: 'Cancelled' }
+          )
         end
       end
